move filter parse methods
