+++ src/server/endpoints/covidcast.py
@@ -22,6 +22,9 @@ from .._params import (
     parse_single_source_signal_arg,
     parse_single_time_arg,
     parse_single_geo_arg,
+    parse_geo_filters,
+    parse_source_signal_filters,
+    parse_time_filters,
 )
 from .._query import QueryBuilder, execute_query, run_query, parse_row, filter_fields
 from .._printer import create_printer, CSVPrinter
@@ -44,51 +47,6 @@ alias = None
 
 latest_table = "epimetric_latest_v"
 history_table = "epimetric_full_v"
-
-def parse_source_signal_filters() -> List[SourceSignalFilter]:
-    ds = request.values.get("data_source")
-    if ds:
-        # old version
-        require_any("signal", "signals", empty=True)
-        signals = extract_strings(("signals", "signal"))
-        if len(signals) == 1 and signals[0] == "*":
-            return [SourceSignalFilter(ds, True)]
-        return [SourceSignalFilter(ds, signals)]
-
-    if ":" not in request.values.get("signal", ""):
-        raise ValidationFailedException("missing parameter: signal or (data_source and signal[s])")
-
-    return parse_source_signal_arg()
-
-
-def parse_geo_filters() -> List[GeoFilter]:
-    geo_type = request.values.get("geo_type")
-    if geo_type:
-        # old version
-        require_any("geo_value", "geo_values", empty=True)
-        geo_values = extract_strings(("geo_values", "geo_value"))
-        if len(geo_values) == 1 and geo_values[0] == "*":
-            return [GeoFilter(geo_type, True)]
-        return [GeoFilter(geo_type, geo_values)]
-
-    if ":" not in request.values.get("geo", ""):
-        raise ValidationFailedException("missing parameter: geo or (geo_type and geo_value[s])")
-
-    return parse_geo_arg()
-
-
-def parse_time_filters() -> TimeFilter:
-    time_type = request.values.get("time_type")
-    if time_type:
-        # old version
-        require_all("time_type", "time_values")
-        time_values = extract_dates("time_values")
-        return TimeFilter(time_type, time_values)
-
-    if ":" not in request.values.get("time", ""):
-        raise ValidationFailedException("missing parameter: time or (time_type and time_values)")
-
-    return parse_time_arg()
 
 
 def _apply_lag_filter(q: QueryBuilder, lag: Optional[int]):
